Fixed the distributer, but IDK what happened with the other 2 files
Now I need to change the default optimization steps and the default reoptimization steps to 5000

I also need to let the user add their own custom  food item to a meal. The optimizer cant touch it because it will stay where the user put it.

Aditionally, the user should be able to add a meal from the inventory

Also, they should be able to remove a food item and put it back in the inventory

With all of this, whenever the user changes something about the diet, the information about the mean diet macros will automatically update